added comment ADmin controller

+++ src/main/java/com/app/service/AdminService.java
@@ -22,7 +22,7 @@ public class AdminService {
 		return ar.findAll();
 	}
 	
-	public Optional<Admin> findById(Long adminId){
+	public Optional<Admin> findById(int adminId){
 		return ar.findById(adminId);
 	}
